+++ commitlint.config.js
@@ -54,14 +54,13 @@ module.exports = {
             emoji: '📚'
           },
           style: {
-            description:
-                            'Changes that do not affect the meaning of the code (white-space, formatting, missing semi-colons, etc)',
+            description: `Changes that do not affect the meaning of the code 
+                          (white-space, formatting, missing semi-colons, etc)`,
             title: 'Styles',
             emoji: '💎'
           },
           refactor: {
-            description:
-                            'A code change that neither fixes a bug nor adds a feature',
+            description: 'A code change that neither fixes a bug nor adds a feature',
             title: 'Code Refactoring',
             emoji: '📦'
           },
@@ -71,26 +70,24 @@ module.exports = {
             emoji: '🚀'
           },
           test: {
-            description:
-                            'Adding missing tests or correcting existing tests',
+            description: 'Adding missing tests or correcting existing tests',
             title: 'Tests',
             emoji: '🚨'
           },
           build: {
-            description:
-                            'Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)',
+            description: `Changes that affect the build system or external dependencies 
+                          (example scopes: gulp, broccoli, npm)`,
             title: 'Builds',
             emoji: '🛠'
           },
           ci: {
-            description:
-                            'Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)',
+            description: `Changes to our CI configuration files and scripts 
+                          (example scopes: Travis, Circle, BrowserStack, SauceLabs)`,
             title: 'Continuous Integrations',
             emoji: '⚙️'
           },
           chore: {
-            description:
-                            'Other changes that don\'t modify src or test files',
+            description: 'Other changes that don\'t modify src or test files',
             title: 'Chores',
             emoji: '♻️'
           },
@@ -102,12 +99,10 @@ module.exports = {
         }
       },
       scope: {
-        description:
-                    'What is the scope of this change (e.g. component or file name)'
+        description: 'What is the scope of this change (e.g. component or file name)'
       },
       subject: {
-        description:
-                    'Write a short, imperative tense description of the change'
+        description: 'Write a short, imperative tense description of the change'
       },
       body: {
         description: 'Provide a longer description of the change'
@@ -116,8 +111,8 @@ module.exports = {
         description: 'Are there any breaking changes?'
       },
       breakingBody: {
-        description:
-                    'A BREAKING CHANGE commit requires a body. Please enter a longer description of the commit itself'
+        description: `A BREAKING CHANGE commit requires a body. 
+                      Please enter a longer description of the commit itself`
       },
       breaking: {
         description: 'Describe the breaking changes'
@@ -126,12 +121,11 @@ module.exports = {
         description: 'Does this change affect any open issues?'
       },
       issuesBody: {
-        description:
-                    'If issues are closed, the commit requires a body. Please enter a longer description of the commit itself'
+        description: `If issues are closed, the commit requires a body. 
+                      Please enter a longer description of the commit itself`
       },
       issues: {
-        description:
-                    'Add issue references (e.g. "fix #123", "re #123".)'
+        description: 'Add issue references (e.g. "fix #123", "re #123".)'
       }
     }
   }