+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [UnReleased] - 2022.07.27
+
+### changed
+
+- Remove InternalError error as much as possible (don't effect using).
+- Change `HashSet` api `eraseElementByValue`'s name to `eraseElementByKey`.
+
 ## [4.0.0-beta.0] - 2022-07-24
 
 ### Added